test(eip7702) add test cases
